public_interface/requirements.txt
# TradeOracle Nexus - Dependencies
streamlit>=1.32.0
langchain>=0.3.0
langchain-google-genai>=3.1.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-elasticsearch>=0.3.0
google-generativeai>=0.8.0
elasticsearch>=8.17.0
ccxt>=4.0.0
numpy>=1.24.0
plotly>=5.18.0
python-dotenv>=1.0.0
pandas>=2.0.0
mcp>=1.26.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
